ieee802154 mac verison for si4463. 1. notice that the GPIO2&3 for rf4463f30 are GPIO2:RX_STATE； GPIO3:TX_STATE which are different to XL4463. 2. The RF_GPIO_CFG are weird. The GPIO0 always 0x1B in the WDS generated .h file which can influence the TX somehow.

+++ kernel_driver/si4463/si4463_api.h
@@ -50,14 +50,14 @@ void enable_chip_irq(void);
 void tx_start(void);
 void rx_start(void);					// 开始接收
 void tx_start_1B(void);
-void tx_set_packet_len(u16 packetlen);
+void tx_set_packet_len(u8 packetlen);
 
 void change_state2tx_tune(void);
 
 void set_frr_ctl(void);
 //Status Getting Method
 void request_device_state(void);
-u16 get_packet_info(void);
+u8 get_packet_info(void);
 void get_ph_status(u8 *rx);
 void get_modem_status(u8 *rx);
 void get_fifo_info(u8 * rx);
